+++ Capstone/FarmStore3/FarmServices/IFarmService.cs
@@ -10,10 +10,9 @@ namespace FarmStore3.FarmServices
 {
     public interface IFarmService
     {
-        ProductViewModel GetProducts();
-
+        ProductViewModel GetAllProducts();
+        Products GetProduct(int produceID);
         ProductViewModel AddNewProduct(Products model);
-
         ProductViewModel UpdateProduce(int id, Products produce);
     }
 
@@ -26,22 +25,23 @@ namespace FarmStore3.FarmServices
             _farmStore = farmStore;
         }
 
-        public ProductViewModel GetProducts()
+        public ProductViewModel GetAllProducts()
         {
-            var dalProducts = _farmStore.SelectAllProduct();
-            var products = new List<Products>();
+            var dalProducts = _farmStore.SelectAllProducts();
+            return MapProductsViewModel(dalProducts);
+            //var products = new List<Products>();
 
-            foreach (var dalProduct in dalProducts)
-            {
-                var product = new Products();
-                product.ProduceName = dalProduct.ProduceName;
-                product.ProduceID = dalProduct.ProduceID;
-                products.Add(product);
-            }
-            var ProductViewModel = new ProductViewModel();
-            ProductViewModel.Products = products;
+            //foreach (var dalProduct in dalProducts)
+            //{
+            //    var product = new Products();
+            //    product.ProduceName = dalProduct.ProduceName;
+            //    product.ProduceID = dalProduct.ProduceID;
+            //    products.Add(product);
+            //}
+            //var ProductViewModel = new ProductViewModel();
+            //ProductViewModel.product = products;
 
-            return ProductViewModel;
+            //return ProductViewModel;
         }
 
         public ProductViewModel AddNewProduct(Products model)
@@ -51,7 +51,7 @@ namespace FarmStore3.FarmServices
             _farmStore.InsertNewProduct(dalModel);
 
             //MAPPING
-            var dalProducts = _farmStore.SelectAllProduct();
+            var dalProducts = _farmStore.SelectAllProducts();
             var products = new List<Products>();
 
             foreach (var dalProduct in dalProducts)
@@ -70,31 +70,53 @@ namespace FarmStore3.FarmServices
         public ProductViewModel UpdateProduce(int id, Products produce)
         {
             _farmStore.UpdateProduct(id, produce);
-            var dalProducts = _farmStore.SelectAllProduct();
+            var dalProducts = _farmStore.SelectAllProducts();
 
             return MapProductViewModel(dalProducts);
         }
 
-        private ProductViewModel MapProductViewModel(IEnumerable<FarmDALModel> dalProducts)
+        public Products GetProduct(int produceID)
         {
-            var _products = new List<Products>();
-            foreach (FarmDALModel item in dalProducts)
+            var dalProduct = _farmStore.SelectProduceByProduceID(produceID);
+            var product = MapProductViewModel(dalProduct);
+            return product;
+        }
+
+        private Products MapProductViewModel(FarmDALModel dalProduct)
+        {
+            var product = new Products();
+            product.ProduceName = dalProduct.ProduceName;
+            product.ProduceID = dalProduct.ProduceID;
+            product.UnitPrice = dalProduct.UnitPrice;
+            product.StockQuantity = dalProduct.StockQuantity;
+            product.InSeason = dalProduct.InSeason;
+            return product;
+        }
+
+        private FarmDALModel MapToFarmDalModel(AddProductViewModel src)
+        {
+            var dalProduct = new FarmDALModel();
+            dalProduct.ProduceName = src.ProduceName;
+            dalProduct.ProduceID = src.ProduceID;
+            dalProduct.StockQuantity = src.StockQuantity;
+            dalProduct.UnitPrice = src.UnitPrice;
+            dalProduct.InSeason = src.InSeason;
+            return dalProduct;
+        }
+
+        private ProductViewModel MapProductsViewModel(IEnumerable<FarmDALModel> dalProducts)
+        {
+            var products = new List<ProductListItemViewModel>();
+
+            foreach (var dalProduct in dalProducts)
             {
-                var product = new Products();
-
-                product.ProduceID = item.ProduceID;
-                product.ProduceName = item.ProduceName;
-                product.StockQuantity = item.StockQuantity;
-                product.CartQuantity = item.CartQuantity;
-                product.UnitPrice = item.UnitPrice;
-                product.InSeason = item.InSeason;
-
-                _products.Add(product);
+                products.Add(new ProductListItemViewModel(dalProduct));
             }
 
-            var model = new ProductViewModel();
-            model.Products = _products;
-            return model;
+            var productViewModel = new ProductViewModel();
+            productViewModel.ListOfProducts = products;
+
+            return productViewModel;
         }
 
         public ProductViewModel DeletAProduct(Products model)
@@ -104,7 +126,7 @@ namespace FarmStore3.FarmServices
             _farmStore.DeleteTheProduct(dalModel);
 
             //MAPPING
-            var dalProducts = _farmStore.SelectAllProduct();
+            var dalProducts = _farmStore.SelectAllProducts();
             var products = new List<Products>();
 
             foreach (var dalProduct in dalProducts)
