+++ Capstone/FarmStore3/Models/ProductViewModel.cs
@@ -4,7 +4,7 @@ namespace FarmStore3.Models
 {
     public class ProductViewModel
     {
-        //public IEnumerable<Products> Products { get; set; }
+        public IEnumerable<Products> Products { get; set; }
         public IEnumerable<ProductListItemViewModel> ListOfProducts { get; set; }
 
     }
